refactor(fs.stat): reorganize code
