+++ packages/fs/fs.stat/src/settings.ts
@@ -7,7 +7,7 @@ export interface Options {
 	throwErrorOnBrokenSymbolicLink?: boolean;
 }
 
-export default class Settings {
+export class Settings {
 	public readonly followSymbolicLink: boolean;
 	public readonly fs: fs.FileSystemAdapter;
 	public readonly markSymbolicLink: boolean;
