+++ packages/fs/fs.stat/src/types/index.ts
@@ -2,3 +2,4 @@ import type * as fs from 'fs';
 
 export type Stats = fs.Stats;
 export type ErrnoException = NodeJS.ErrnoException;
+export type AsyncCallback = (error: ErrnoException | null, stats: Stats) => void;
